# docs/openapi-v1.yaml
openapi: 3.1.0
info:
  title: SubTerrat API
  version: 0.1.0
  description: >-
    Contract-first API for frozen Taipei citizen-reported rat-hotspot spatial
    rankings. This specification does not assert that an implementation or a
    validated prediction model exists.
  x-implementation-status: SPECIFICATION_ONLY
  x-evidence-status: NO_TRUSTED_RESULT
...
servers:
  - url: http://localhost:8080
    description: Local development placeholder; no runtime is asserted.
...
paths:
...
  /api/v1/releases/{release_id}/cells:
    parameters:
      - $ref: "#/components/parameters/ReleaseId"
      - $ref: "#/components/parameters/Bbox"
      - $ref: "#/components/parameters/Layers"
      - $ref: "#/components/parameters/Limit"
      - $ref: "#/components/parameters/PageToken"
    get:
      tags: [Map, Predictions]
      operationId: listReleaseCells
      summary: Read frozen S2 cells intersecting a WGS84 bounding box
      description: >-
        Returns only sanitized public serving fields. The response never reads
        exact T1 report points, raw reports, complete sewer geometry, or precise
        residences and businesses.
      security: []
      responses:
        "200":
          description: GeoJSON feature collection of frozen S2 cells
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
          content:
            application/geo+json:
              schema:
                $ref: "#/components/schemas/CellFeatureCollection"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/v1/releases/{release_id}/cells/{cell_id}:
    parameters:
      - $ref: "#/components/parameters/ReleaseId"
      - $ref: "#/components/parameters/CellId"
    get:
      tags: [Map, Predictions]
      operationId: getReleaseCell
      summary: Read one frozen cell and its explanation fields
      security: []
      responses:
        "200":
          description: Cell detail
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
          content:
            application/geo+json:
              schema:
                $ref: "#/components/schemas/CellFeature"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    googleOidc:
      type: openIdConnect
      openIdConnectUrl: https://accounts.google.com/.well-known/openid-configuration
      description: Internal Cloud Run IAM/OIDC identity; never embedded as a Flutter secret.
  headers:
    ETag:
      description: Validator derived from the immutable frozen payload digest.
      schema:
        type: string
  parameters:
    ReleaseId:
      name: release_id
      in: path
      required: true
      schema:
        type: string
        minLength: 1
...
    CellId:
      name: cell_id
      in: path
      required: true
      description: S2 cell identifier encoded as a string to preserve 64-bit precision.
      schema:
        type: string
        pattern: "^[0-9]+$"
    Bbox:
      name: bbox
      in: query
      required: true
      description: WGS84 west,south,east,north bounding box.
      schema:
        type: string
        pattern: "^-?[0-9]+(?:\\.[0-9]+)?,-?[0-9]+(?:\\.[0-9]+)?,-?[0-9]+(?:\\.[0-9]+)?,-?[0-9]+(?:\\.[0-9]+)?$"
    Layers:
      name: layers
      in: query
      required: false
      style: form
      explode: false
      schema:
        type: array
        minItems: 1
        uniqueItems: true
        items:
          type: string
          enum: [structural_score, data_coverage, approved_report_count]
        default: [structural_score, data_coverage]
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 1500
        default: 500
    PageToken:
      name: page_token
      in: query
      required: false
      schema:
        type: string
        minLength: 1
...
  responses:
...
    NotFound:
      description: Requested immutable resource was not found
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    ValidationError:
      description: Request violates the declared API contract
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
  schemas:
...
    ForecastWindow:
      type: object
      additionalProperties: false
      required: [start, end]
      properties:
        start:
          type: string
          format: date-time
        end:
          type: string
          format: date-time
...
    ScoreComponents:
      type: object
      additionalProperties: false
      required: [food, sewer, abandoned]
      properties:
        food:
          type: [number, "null"]
          minimum: 0
          maximum: 1
        sewer:
          type: [number, "null"]
          minimum: 0
          maximum: 1
        abandoned:
          type: [number, "null"]
          minimum: 0
          maximum: 1
    RankVariants:
      type: object
      additionalProperties: false
      required:
        - main
        - food_only
        - without_food
        - without_sewer
        - without_abandoned
      properties:
        main:
          type: [integer, "null"]
          minimum: 1
        food_only:
          type: [integer, "null"]
          minimum: 1
        without_food:
          type: [integer, "null"]
          minimum: 1
        without_sewer:
          type: [integer, "null"]
          minimum: 1
        without_abandoned:
          type: [integer, "null"]
          minimum: 1
    TopKFlags:
      type: object
      additionalProperties: false
      required: [top_05, top_10, top_20]
      properties:
        top_05:
          type: boolean
        top_10:
          type: boolean
        top_20:
          type: boolean
    ReleasedOutcomeSummary:
      type: object
      additionalProperties: false
      required: [availability, suppressed, approved_report_count, count_band]
      properties:
        availability:
          type: string
          enum: [released, withheld, not_observed]
        suppressed:
          type: boolean
        approved_report_count:
          type: [integer, "null"]
          minimum: 0
        count_band:
          type: [string, "null"]
        observation_window:
          oneOf:
            - $ref: "#/components/schemas/ForecastWindow"
            - type: "null"
        field_verified:
          type: boolean
          const: false
    RawLayerFields:
      description: >-
        Passthrough of `map_hotspot_cells_t0` columns (see
        sql/structural/07_materialize_map_payload.sql) that don't map onto
        any modeled CellProperties field, kept as its own labeled object
        instead of loosening CellProperties' additionalProperties. Added for
        GitHub issue #4's Cloud Run MVP, which serves this table directly.
      type: object
      additionalProperties: false
      properties:
        food_top_area:
          type: [boolean, "null"]
        sewer_top_area:
          type: [boolean, "null"]
        sewer_has_score:
          type: [boolean, "null"]
        sewer_coverage_state:
          type: [string, "null"]
        unused_public_building_address_point_count:
          type: [integer, "null"]
          minimum: 0
        building_overlay_state:
          type: [string, "null"]
        freeze_id:
          type: [string, "null"]
        model_kind:
          type: [string, "null"]
        score_semantics:
          type: [string, "null"]
        evidence_state:
          type: [string, "null"]
    CellProperties:
      type: object
      additionalProperties: false
      required:
        - cell_id
        - prediction_run_id
        - score_kind
        - structural_score
        - rank_percentile
        - calibrated_probability
        - components
        - ranks
        - top_k
        - coverage_state
        - freshness
        - limitation_codes
      properties:
        cell_id:
          type: string
          pattern: "^[0-9]+$"
        prediction_run_id:
          type: string
        score_kind:
          type: string
          const: structural_report_hotspot_ranking
        structural_score:
          type: [number, "null"]
          minimum: 0
          maximum: 1
        rank_percentile:
          type: [number, "null"]
          minimum: 0
          maximum: 1
        calibrated_probability:
          type: "null"
        components:
          $ref: "#/components/schemas/ScoreComponents"
        ranks:
          $ref: "#/components/schemas/RankVariants"
        top_k:
          # null when no combined three-group MainScore has been computed
          # yet (e.g. the abandoned-building layer is overlay-only, not a
          # ranked citywide feature — see contracts/structural_score_v0_1.json
          # layerwise_policy.missing_layer_action). Never fabricated.
          oneOf:
            - $ref: "#/components/schemas/TopKFlags"
            - type: "null"
        coverage_state:
          type: string
          enum: [eligible, unscored_missing_group, excluded_boundary, withheld]
        freshness:
          type: string
          enum: [current, stale, unknown]
        limitation_codes:
          type: array
          minItems: 1
          items:
            type: string
        released_outcome:
          $ref: "#/components/schemas/ReleasedOutcomeSummary"
        raw_layer_fields:
          $ref: "#/components/schemas/RawLayerFields"
    Position:
      type: array
      prefixItems:
        - type: number
          minimum: -180
          maximum: 180
        - type: number
          minimum: -90
          maximum: 90
      minItems: 2
      maxItems: 2
    LinearRing:
      type: array
      minItems: 4
      items:
        $ref: "#/components/schemas/Position"
    PolygonCoordinates:
      type: array
      minItems: 1
      items:
        $ref: "#/components/schemas/LinearRing"
    GeoJsonPolygon:
      type: object
      additionalProperties: false
      required: [type, coordinates]
      properties:
        type:
          type: string
          const: Polygon
        coordinates:
          $ref: "#/components/schemas/PolygonCoordinates"
    GeoJsonMultiPolygon:
      type: object
      additionalProperties: false
      required: [type, coordinates]
      properties:
        type:
          type: string
          const: MultiPolygon
        coordinates:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/PolygonCoordinates"
    CellFeature:
      type: object
      additionalProperties: false
      required: [type, id, geometry, properties]
      properties:
        type:
          type: string
          const: Feature
        id:
          type: string
          pattern: "^[0-9]+$"
        geometry:
          oneOf:
            - $ref: "#/components/schemas/GeoJsonPolygon"
            - $ref: "#/components/schemas/GeoJsonMultiPolygon"
        properties:
          $ref: "#/components/schemas/CellProperties"
    CellFeatureCollection:
      type: object
      additionalProperties: false
      required:
        - type
        - release_id
        - prediction_run_id
        - target_window
        - features
        - next_page_token
        - truncated
      properties:
        type:
          type: string
          const: FeatureCollection
        release_id:
          type: string
        prediction_run_id:
          type: string
        target_window:
          # null for a development-exposed retrospective freeze that has no
          # real official_t0 observation window (docs/V0_1_RUNBOOK.md) —
          # never fabricated.
          oneOf:
            - $ref: "#/components/schemas/ForecastWindow"
            - type: "null"
        features:
          type: array
          maxItems: 1500
          items:
            $ref: "#/components/schemas/CellFeature"
        next_page_token:
          type: [string, "null"]
        truncated:
          type: boolean
          const: false
...
    Problem:
      type: object
      additionalProperties: false
      required: [type, title, status, code, detail, request_id]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        code:
          type: string
          enum:
            - NO_PUBLISHED_RELEASE
            - RELEASE_NOT_FOUND
            - CELL_NOT_FOUND
            - PREDICTION_RUN_NOT_FOUND
            - INVALID_BBOX
            - INVALID_TIME_WINDOW
            - CAPABILITY_NOT_AVAILABLE
            - DATA_GATE_FAILED
            - STATE_CONFLICT
            - IDEMPOTENCY_CONFLICT
            - FORBIDDEN_DATA_SCOPE
            - INTERNAL_ERROR
        detail:
          type: string
        request_id:
          type: string